fix(powl,powl-receipt,mcp): clear all clippy -D warnings debt in the 5 MFW retrofit crates
make clippy's own [tasks.clippy] never actually reaches any MFW-retrofit
code: its lint-anti-llm dependency shells out to the external
/Users/sac/anti-llm-cheat-lsp repo, which fails to compile (affidavit
26.6.22's model_mining.rs:175 calls PetriNet::new on a
wasm4pm_compat::petri::Marking that doesn't implement
IntoIterator<Item=(String,usize)>, per wasm4pm-core/src/models.rs:1039 --
confirmed via a direct `cargo build --manifest-path
/Users/sac/anti-llm-cheat-lsp/Cargo.toml`). That chain is entirely external
(a separate repo, a crates.io dependency, and another separate repo) and
out of this fix's scope -- BLOCKED, not something this commit resolves.

Running clippy directly instead (bypassing lint-anti-llm) leaves genuine
UNVERIFIED status for the MFW crates' own cleanliness, since
`--workspace --all-targets` halts on bcinr-logic's large, pre-existing,
non-MFW test-code lint debt (49 errors under --all-targets -- unaffected by
this commit, confirmed via `cargo clippy -p bcinr-logic --all-targets --
-D warnings`, none in parse.rs) before ever reaching any MFW crate.

This commit closes that verification gap the other way: run clippy -D
warnings against each of the 5 MFW retrofit crates individually
(bcinr-mfw-ir, bcinr-pddl --features mfw-planner, bcinr-powl,
bcinr-powl-receipt, bcinr-mcp), all --all-targets, and fix every finding
that was actually theirs (9 total, all pre-existing, none introduced by
this session's other fixes):

- bcinr-powl: needless_range_loop x2 (compiler.rs build_in_degrees --
  #[allow] with a comment, since `i` indexes two different collections and
  clippy's enumerate() rewrite doesn't apply cleanly; scheduler_wide.rs
  Step 3 -- real fix via iter_mut().zip()), let_unit_value (dispatcher.rs's
  deliberate const-assert-forcing statement -- #[allow] with a comment,
  since removing the `let` just trades this lint for rustc's own
  path_statements), len_without_is_empty (receipt_worker.rs ReceiptLog --
  added is_empty()), unnecessary_cast x2 (scheduler.rs, tape.rs),
  new_without_default (scheduler_wired.rs FiberPool -- added Default impl
  delegating to the existing const fn new()).
- bcinr-powl-receipt: derivable_impls (causal_receipt.rs PackedObjRef's
  manual Default -> #[derive(Default)], u32::default()==0 matches exactly),
  needless_range_loop (ocel_emit.rs -- real fix via iter_mut().zip(), zip
  naturally bounds iteration to min(8, obj_refs.len()) without a separate
  index).
- bcinr-mcp: manual_strip (adversarial.rs test), option_map_unit_fn +
  len_zero (integration_tests.rs, both test-only).

Result: `cargo clippy -p <crate> --all-targets -- -D warnings` is now clean
for all 5 MFW crates individually -- ALIVE, not merely "blocked by the
external tool" as item 26 characterized it before this fix.

cargo test --lib (and --all-targets where applicable) for bcinr-powl
(220/220), bcinr-powl-receipt (62+8+5/all pass), bcinr-mcp (23+12/all
pass): no regressions from any of the 9 mechanical fixes. cargo fmt --check
clean for all three crates.

+++ crates/bcinr-powl/src/tape.rs
@@ -223,7 +223,7 @@ pub mod v2 {
     #[inline(always)]
     pub const fn eq_mask_u64(a: u64, b: u64) -> u64 {
         let diff = a ^ b;
-        let nonzero_bit = ((diff | diff.wrapping_neg()) >> 63) as u64;
+        let nonzero_bit = (diff | diff.wrapping_neg()) >> 63;
         nonzero_bit.wrapping_sub(1)
     }
